log_simul/requirements.txt
Faker==28.4.1
ulid-py==1.1.0
httpx==0.27.2
python-dotenv==1.0.1
backports.zoneinfo==0.2.1; python_version < "3.9"

# Kafka client for Python (used by output_plugins.kafka_output)
kafka-python==2.0.2

